intrusionx-se/requirements.txt
gradio>=4.0
torch
torchvision
transformers
Pillow
opencv-python-headless
librosa
soundfile
exifread
matplotlib
numpy
scipy
fastapi
uvicorn[standard]
python-multipart
reportlab
